Consolidate RT memory setup into rt_memory
Split rt_memory into management (locking/mallopt/reserve, in rt_memory.hpp/.cpp)
and reporting (the periodic page-fault/memory-footprint monitor, in
rt_memory_reporting.hpp/.cpp) -- rt_memory.cpp had grown to hold both, per
review feedback.

Make configure_realtime_memory() ROS-free: it now returns a plain
RealtimeMemorySetup struct (rlimit values, mlockall success/errno) instead of
taking an rclcpp::Logger and logging itself. control_node.cpp does its own
logging from the result. This drops rt_memory.cpp's rclcpp dependency entirely,
which is what lets test/manual/mlock_demo.cpp link it directly with a plain
compiler (next commit).

Split the unit tests to match (test_rt_memory.cpp / test_rt_memory_reporting.cpp)
and wire the new source/test target into CMakeLists.txt.

Refs SW-933
