+++ eli_cs_robot_driver/CMakeLists.txt
@@ -164,6 +164,7 @@ add_executable(
   eli_ros2_control_node
   src/control_node.cpp
   src/rt_memory.cpp
+  src/rt_memory_reporting.cpp
 )
 ament_target_dependencies(
   eli_ros2_control_node
@@ -223,10 +224,14 @@ if(BUILD_TESTING)
   set(ament_cmake_cpplint_FOUND TRUE)
   ament_lint_auto_find_test_dependencies()
 
-  # Unit tests for the ROS-free real-time memory helpers (rt_memory.hpp).
+  # Unit tests for the ROS-free real-time memory MANAGEMENT helpers (rt_memory.hpp).
   find_package(ament_cmake_gtest REQUIRED)
   ament_add_gtest(test_rt_memory test/test_rt_memory.cpp)
   target_include_directories(test_rt_memory PRIVATE include)
+
+  # Unit tests for the real-time memory REPORTING helpers (rt_memory_reporting.hpp).
+  ament_add_gtest(test_rt_memory_reporting test/test_rt_memory_reporting.cpp)
+  target_include_directories(test_rt_memory_reporting PRIVATE include)
 endif()
 
 ament_package()
